refactor(toast-util): toast-util with functions to be used in both ToastController and webcomponent
