+++ packages/crayons-core/src/components/toast-child/toast-child.stories.mdx
@@ -1,0 +1,57 @@
+import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
+
+# ToastChild
+
+<Meta title='Toast' component='fw-toast-child' />
+
+## Props
+
+<Props of='fw-toast-child' />
+
+## Usage
+
+### Default
+
+<Preview>
+  <Story name='Default'>
+    {() => `
+    <fw-toast-child id="default_toast" type="success" content="Successfullly triggered" sticky="true"></fw-toast-child>
+    `}
+  </Story>
+</Preview>
+
+### Type
+
+<Preview>
+  <Story name='Type'>
+    {() => `
+    <fw-toast-child id="type_toast" sticky=true type="success" content="success"></fw-toast-child>
+ <fw-toast-child id="type_toast" sticky=true type="error" content="error"></fw-toast-child>
+ <fw-toast-child id="type_toast" sticky=true type="warning" content="warning"></fw-toast-child>
+ <fw-toast-child id="type_toast" sticky=true type="inprogress" content="inprogress"></fw-toast-child>
+    `}
+  </Story>
+</Preview>
+
+### Position
+
+<Preview>
+  <Story name='position'>
+    {() => `
+    <div style="margin-top: 100px">
+      <fw-toast-child id="topleft" type="success" position="top-left" content="Successfullly triggered"></fw-toast-child>
+      <fw-toast-child id="topcenter" type="success" position="top-center" content="Successfullly triggered"></fw-toast-child>
+      <fw-toast-child id="topright" type="success" position="top-right" content="Successfullly triggered"></fw-toast-child>
+    </div>
+    `}
+  </Story>
+</Preview>
+
+### Action Link
+
+<Preview>
+  <Story name='Action Link'>
+    {() => `
+    <fw-toast-child id="actionlink_toast" action-link-text="Undo" type="success" content="Record successfullly deleted" sticky="true"></fw-toast-child>`}
+  </Story>
+</Preview>
