--- packages/crayons-core/src/components/toast-controller/readme.md
@@ -1,46 +1,0 @@
-# fw-toast-controller
-
-
-
-<!-- Auto Generated Below -->
-
-
-## Properties
-
-| Property   | Attribute  | Description                                  | Type                                        | Default        |
-| ---------- | ---------- | -------------------------------------------- | ------------------------------------------- | -------------- |
-| `position` | `position` | position of the toast notification in screen | `"top-center" \| "top-left" \| "top-right"` | `'top-center'` |
-| `timeout`  | `timeout`  | Time duration of the toast visibility        | `number`                                    | `4000`         |
-
-
-## Methods
-
-### `trigger(opts: any) => Promise<void>`
-
-
-
-#### Returns
-
-Type: `Promise<void>`
-
-
-
-
-## Dependencies
-
-### Depends on
-
-- [fw-toast](../toast)
-
-### Graph
-```mermaid
-graph TD;
-  fw-toast-controller --> fw-toast
-  fw-toast --> fw-spinner
-  fw-toast --> fw-icon
-  style fw-toast-controller fill:#f9f,stroke:#333,stroke-width:4px
-```
-
-----------------------------------------------
-
-Built with ❤ at Freshworks
