--- packages/crayons-core/src/components/toast-controller/toast-controller.stories.mdx
@@ -1,24 +1,0 @@
-import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
-
-# ToastController
-
-<Meta
-  title='ToastController'
-  component='fw-toast-controller'
-/>
-
-## Props 
-
-<Props of="fw-toast-controller"/>
-
-## Usage
-
-### Default
-<Preview>
-  <Story name='Default'>
-    {
-      () => `<fw-toast-controller></fw-toast-controller>`
-    }
-  </Story>
-</Preview>
-
