+++ packages/crayons-core/src/components/toast/readme.md
@@ -6,10 +6,10 @@ Toasts are used to show pop-up messages that lasts on the screen for a while. Us
 
 ```html live
 <fw-toast id="type_toast"></fw-toast>
-<fw-button onClick="document.querySelector('#type_toast').trigger({type:'success', content: 'Successfullly triggered'})">Success</fw-button>
-<fw-button onClick="document.querySelector('#type_toast').trigger({type:'error', content:'something went wrong!'})">Error</fw-button>
-<fw-button onClick="document.querySelector('#type_toast').trigger({type:'warning', content:'This is a warning!'})">Warning</fw-button>
-<fw-button onClick="document.querySelector('#type_toast').trigger({type:'inprogress', content:'Request is in progress'})">Inprogress</fw-button>
+<fw-button onclick="document.querySelector('#type_toast').trigger({type:'success', content: 'Successfullly triggered'})">Success</fw-button>
+<fw-button onclick="document.querySelector('#type_toast').trigger({type:'error', content:'something went wrong!'})">Error</fw-button>
+<fw-button onclick="document.querySelector('#type_toast').trigger({type:'warning', content:'This is a warning!'})">Warning</fw-button>
+<fw-button onclick="document.querySelector('#type_toast').trigger({type:'inprogress', content:'Request is in progress'})">Inprogress</fw-button>
 ```
 
 <!-- Auto Generated Below -->
@@ -17,41 +17,42 @@ Toasts are used to show pop-up messages that lasts on the screen for a while. Us
 
 ## Properties
 
-| Property         | Attribute          | Description                                              | Type                                                | Default     |
-| ---------------- | ------------------ | -------------------------------------------------------- | --------------------------------------------------- | ----------- |
-| `actionLinkText` | `action-link-text` | The Content of the action link                           | `string`                                            | `''`        |
-| `content`        | `content`          | The content to be diaplyed in toast                      | `string`                                            | `undefined` |
-| `pauseOnHover`   | `pause-on-hover`   | Pause the toast from hiding on mouse hover               | `boolean`                                           | `undefined` |
-| `sticky`         | `sticky`           | won't close automatically                                | `boolean`                                           | `false`     |
-| `timeout`        | `timeout`          | Time duration of the toast visibility                    | `number`                                            | `4000`      |
-| `type`           | `type`             | Type of the toast - success,failure, warning, inprogress | `"error" \| "inprogress" \| "success" \| "warning"` | `'warning'` |
+| Property         | Attribute          | Description                                              | Type                                                | Default        |
+| ---------------- | ------------------ | -------------------------------------------------------- | --------------------------------------------------- | -------------- |
+| `actionLinkText` | `action-link-text` | The Content of the action link                           | `string`                                            | `''`           |
+| `content`        | `content`          | The content to be displayed in toast                     | `string`                                            | `undefined`    |
+| `pauseOnHover`   | `pause-on-hover`   | Pause the toast from hiding on mouse hover               | `boolean`                                           | `undefined`    |
+| `position`       | `position`         | position of the toast notification in screen             | `"top-center" \| "top-left" \| "top-right"`         | `'top-center'` |
+| `sticky`         | `sticky`           | won't close automatically                                | `boolean`                                           | `false`        |
+| `timeout`        | `timeout`          | Time duration of the toast visibility                    | `number`                                            | `4000`         |
+| `type`           | `type`             | Type of the toast - success,failure, warning, inprogress | `"error" \| "inprogress" \| "success" \| "warning"` | `'warning'`    |
 
 
-## Events
+## Methods
 
-| Event           | Description                                                   | Type               |
-| --------------- | ------------------------------------------------------------- | ------------------ |
-| `fwLinkClick`   | Triggered when the action link clicked.                       | `CustomEvent<any>` |
-| `fwRemoveToast` | Remove toast element from the parent on closing toast message | `CustomEvent<any>` |
+### `trigger(opts: ToastOptions) => Promise<void>`
+
+
+
+#### Returns
+
+Type: `Promise<void>`
+
+
 
 
 ## Dependencies
 
-### Used by
-
- - [fw-toast-controller](../toast-controller)
-
 ### Depends on
 
-- [fw-spinner](../spinner)
-- [fw-icon](../icon)
+- [fw-toast-child](../toast-child)
 
 ### Graph
 ```mermaid
 graph TD;
-  fw-toast --> fw-spinner
-  fw-toast --> fw-icon
-  fw-toast-controller --> fw-toast
+  fw-toast --> fw-toast-child
+  fw-toast-child --> fw-spinner
+  fw-toast-child --> fw-icon
   style fw-toast fill:#f9f,stroke:#333,stroke-width:4px
 ```
