+++ packages/crayons-core/src/components/toast/toast.stories.mdx
@@ -1,19 +1,17 @@
 import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
 
-# Toast
+# ToastController
 
-<Meta
-  title='Toast'
-  component='fw-toast'
-/>
+<Meta title='Toast' component='fw-toast' />
 
-## Props 
+## Props
 
-<Props of="fw-toast"/>
+<Props of='fw-toast' />
 
 ## Usage
 
 ### Default
+
 <Preview>
   <Story name='Default'>
     {() => `
@@ -24,6 +22,7 @@ import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
 </Preview>
 
 ### Type
+
 <Preview>
   <Story name='Type'>
     {() => `
@@ -37,6 +36,7 @@ import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
 </Preview>
 
 ### Position
+
 <Preview>
   <Story name='position'>
     {() => `
@@ -53,6 +53,7 @@ import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
 </Preview>
 
 ### Action Link
+
 <Preview>
   <Story name='Action Link'>
     {() => `
